Refactor tabContent into hook and add to credentials page
