+++ frontend/src/admin/hooks/useCreatorTabs.tsx
@@ -1,0 +1,45 @@
+import { PhotoIcon, CreditCardIcon } from '@heroicons/react/24/outline'
+import { useNavigate } from 'react-router-dom'
+
+import { baseRoute } from '../../client/api/BaseUrl'
+
+interface UseCreatorTabsProps {
+  activeTab: 'showcases' | 'credentials'
+  onTabChange: (tab: 'showcases' | 'credentials') => void
+}
+
+export function useCreatorTabs({ activeTab, onTabChange }: UseCreatorTabsProps) {
+  const navigate = useNavigate()
+
+  const tabsContent = (
+    <div className="flex gap-8">
+      <button
+        onClick={() => {
+          onTabChange('showcases')
+          navigate(`${baseRoute}/admin/creator`)
+        }}
+        className={`pb-2 font-medium transition-colors border-b-2 flex items-center gap-2 ${
+          activeTab === 'showcases'
+            ? 'border-bcgov-blue text-bcgov-blue'
+            : 'border-transparent text-bcgov-darkgrey hover:text-bcgov-black'
+        }`}
+      >
+        <PhotoIcon className="w-5 h-5" />
+        Showcases
+      </button>
+      <button
+        onClick={() => navigate(`${baseRoute}/admin/creator/credentials`)}
+        className={`pb-2 font-medium transition-colors border-b-2 flex items-center gap-2 ${
+          activeTab === 'credentials'
+            ? 'border-bcgov-blue text-bcgov-blue'
+            : 'border-transparent text-bcgov-darkgrey hover:text-bcgov-black'
+        }`}
+      >
+        <CreditCardIcon className="w-5 h-5" />
+        Credentials
+      </button>
+    </div>
+  )
+
+  return tabsContent
+}
